backend/requirements.txt
fastapi
uvicorn
openai
python-dotenv
pyttsx3
SpeechRecognition
pyaudio
